pyproject.toml
[build-system]
requires = ["flit_core >=3.2,<4"]
build-backend = "flit_core.buildapi"

[project]
name = "gpx"
authors = [{name = "Steven van de Graaf", email = "steven@vandegraaf.xyz"}]
readme = "README.md"
license = {file = "LICENSE"}
keywords = ["gpx", "gps", "xml"]
classifiers = [
    "Development Status :: 4 - Beta",
    "License :: OSI Approved :: GNU General Public License v3 or later (GPLv3+)",
    "Programming Language :: Python",
    "Programming Language :: Python :: 3",
    "Programming Language :: Python :: 3.7",
    "Programming Language :: Python :: 3.8",
    "Programming Language :: Python :: 3.9",
    "Programming Language :: Python :: 3.10",
    "Programming Language :: Python :: 3.11",
    "Typing :: Typed",
]
requires-python = "~=3.7"
dependencies = [
    "lxml",
    "python-dateutil",
]
dynamic = ["version", "description"]

[project.optional-dependencies]
dev = [
    "black",
    "mypy",
    "pre-commit",
    "ruff",
]
docs = [
    "furo",
    "myst-parser",
    "sphinx",
    "sphinx-autobuild",
    "sphinx-copybutton",
    "sphinxext-opengraph",
]
tests = [
    "pytest"
]

[project.urls]
Documentation = "https://gpx.readthedocs.io/en/stable/"
Changelog = "https://gpx.readthedocs.io/en/stable/changelog.html"
"Source code" = "https://github.com/sgraaf/gpx"
Issues = "https://github.com/sgraaf/gpx/issues"

[tool.ruff]
select = [
    "B",
    "C90",
    "E",
    "F",
    "I",
    "UP",
    "RUF100",
    "W",
]
ignore = ["E501"]
src = ["src"]
target-version = "py37"

[tool.ruff.per-file-ignores]
"__init__.py" = ["F401"]
